requirements.txt
anthropic>=0.40.0
requests>=2.32.0
slack-sdk>=3.27.0
slack-bolt>=1.18.0
flask>=2.3.0
pydantic>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
